+++ list.c
@@ -1,7 +1,7 @@
 
 /*Created by Omri Kaisari on 05/08/2020.*/
 
-#include<stdlib.h>
+#include <stdlib.h>
 #include "list.h"
 
 
@@ -54,7 +54,7 @@ void list_destroy(List list, DestroyFunction destroy) {
     list_node_destroy(current, destroy);
 }
 
-int list_size(List list) {
+size_t list_size(List list) {
     int size = 0;
     ListNode current = list;
     if (list == NULL)
@@ -100,7 +100,7 @@ size_t list_get_size_of() {
     return sizeof(struct node_t);
 }
 
-void * list_find_element(List list, const void *element, Equals compereFunction) {
+void *list_find_element(List list, const void *element, Equals compereFunction) {
     ListNode current = list;
     while (current != NULL) {
         if (current->data != NULL) {
